pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crds-data-process"
version = "0.1.0"
description = "CRDS数据处理：从衰荡时间到MATS光谱拟合"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "lmfit>=1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "pre-commit>=3.0",
]
notebook = [
    "jupyter>=1.0",
    "ipykernel>=6.0",
]
mats = [
    "MATS-venv",
]

[project.scripts]
crds-process = "crds_process.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

